feat: move update checking into a settings updates tab
Settings gains an Updates tab (current version, automatic-check toggle
backed by Sparkle's persisted preference, manual check button). The
Maintenance popover returns to pure background-helper duty; its version
display and the store's currentVersion go with it.
