+++ Sources/Hostflip/ApplicationSettingsView.swift
@@ -1,7 +1,10 @@
+import HostflipXPC
+import Sparkle
 import SwiftUI
 
 struct ApplicationSettingsView: View {
     let dockIconStore: DockIconVisibilityStore
+    let updater: SPUUpdater
 
     var body: some View {
         TabView {
@@ -26,7 +29,70 @@ struct ApplicationSettingsView: View {
             .tabItem {
                 Label("General", systemImage: "gearshape")
             }
+
+            UpdatesSettingsView(updater: updater)
+                .padding(20)
+                .tabItem {
+                    Label("Updates", systemImage: "arrow.triangle.2.circlepath")
+                }
         }
         .frame(width: 460, height: 150)
+    }
+}
+
+/// The discoverable home of update checking (#34); Sparkle owns the persisted
+/// automatic-check preference, this view only mirrors it.
+private struct UpdatesSettingsView: View {
+    let updater: SPUUpdater
+    @State private var automaticallyChecksForUpdates: Bool
+
+    init(updater: SPUUpdater) {
+        self.updater = updater
+        automaticallyChecksForUpdates = updater.automaticallyChecksForUpdates
+    }
+
+    var body: some View {
+        Form {
+            LabeledContent("Current Version") {
+                Text("v\(HostflipBuild.version)")
+                    .foregroundStyle(.secondary)
+            }
+
+            Toggle("Automatically check for updates", isOn: $automaticallyChecksForUpdates)
+                .onChange(of: automaticallyChecksForUpdates) { _, newValue in
+                    updater.automaticallyChecksForUpdates = newValue
+                }
+
+            CheckForUpdatesButton(updater: updater)
+        }
+    }
+}
+
+/// Sparkle drives the whole check-and-install flow with its own UI; this button only forwards the
+/// click and mirrors `canCheckForUpdates` (false while a check or install is already in flight).
+struct CheckForUpdatesButton: View {
+    @ObservedObject private var viewModel: CheckForUpdatesViewModel
+    private let updater: SPUUpdater
+
+    init(updater: SPUUpdater) {
+        self.updater = updater
+        viewModel = CheckForUpdatesViewModel(updater: updater)
+    }
+
+    var body: some View {
+        Button("Check for Updates") {
+            updater.checkForUpdates()
+        }
+        .disabled(!viewModel.canCheckForUpdates)
+    }
+}
+
+@MainActor
+final class CheckForUpdatesViewModel: ObservableObject {
+    @Published var canCheckForUpdates = false
+
+    init(updater: SPUUpdater) {
+        updater.publisher(for: \.canCheckForUpdates)
+            .assign(to: &$canCheckForUpdates)
     }
 }
