+++ Sources/Hostflip/HelperMaintenanceView.swift
@@ -1,5 +1,4 @@
 import HostflipXPC
-import Sparkle
 import SwiftUI
 
 struct HelperStatusPresentation {
@@ -61,11 +60,11 @@ struct HelperStatusLabel: View {
     }
 }
 
-/// User-facing maintenance panel: the entry point for both helper management and update checks.
+/// User-facing maintenance panel: the entry point for helper management.
+/// Update checks moved to Settings > Updates (#34).
 struct HelperMaintenanceView: View {
     let store: WorkspaceStore
     let maintenanceStore: MaintenanceStore
-    let updater: SPUUpdater
     let onDismiss: (() -> Void)?
     @Environment(\.dismiss) private var dismiss
     @State private var isConfirmingRemoval = false
@@ -73,12 +72,10 @@ struct HelperMaintenanceView: View {
     init(
         store: WorkspaceStore,
         maintenanceStore: MaintenanceStore,
-        updater: SPUUpdater,
         onDismiss: (() -> Void)? = nil
     ) {
         self.store = store
         self.maintenanceStore = maintenanceStore
-        self.updater = updater
         self.onDismiss = onDismiss
     }
 
@@ -119,19 +116,6 @@ struct HelperMaintenanceView: View {
                     }
                     .controlSize(.small)
                 }
-            }
-
-            Divider()
-
-            HStack(spacing: 10) {
-                Label("Updates", systemImage: "arrow.triangle.2.circlepath")
-                    .font(.subheadline.weight(.semibold))
-                Text("v\(maintenanceStore.currentVersion)")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                Spacer()
-                CheckForUpdatesButton(updater: updater)
-                    .controlSize(.small)
             }
 
             if let feedback = maintenanceStore.feedback {
@@ -182,34 +166,5 @@ struct HelperMaintenanceView: View {
         } else {
             dismiss()
         }
-    }
-}
-
-/// Sparkle drives the whole check-and-install flow with its own UI; this button only forwards the
-/// click and mirrors `canCheckForUpdates` (false while a check or install is already in flight).
-struct CheckForUpdatesButton: View {
-    @ObservedObject private var viewModel: CheckForUpdatesViewModel
-    private let updater: SPUUpdater
-
-    init(updater: SPUUpdater) {
-        self.updater = updater
-        viewModel = CheckForUpdatesViewModel(updater: updater)
-    }
-
-    var body: some View {
-        Button("Check for Updates") {
-            updater.checkForUpdates()
-        }
-        .disabled(!viewModel.canCheckForUpdates)
-    }
-}
-
-@MainActor
-final class CheckForUpdatesViewModel: ObservableObject {
-    @Published var canCheckForUpdates = false
-
-    init(updater: SPUUpdater) {
-        updater.publisher(for: \.canCheckForUpdates)
-            .assign(to: &$canCheckForUpdates)
     }
 }
